+++ Editor/Scripts/Attributes/BindDrawerToAttribute.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace RealityProgrammer.OverseerInspector.Editors.Drawers {
+namespace RealityProgrammer.OverseerInspector.Editors.Attributes {
     [AttributeUsage(AttributeTargets.Class, AllowMultiple = true, Inherited = false)]
     public class BindDrawerToAttribute : Attribute {
         public Type AttributeType { get; private set; }