+++ Editor/Scripts/Attributes/ConditionalConnectAttribute.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace RealityProgrammer.OverseerInspector.Editors.Attributes {
+    [AttributeUsage(AttributeTargets.Method, AllowMultiple = false, Inherited = false)]
+    public class ConditionalConnectAttribute : Attribute {
+        public Type ConditionalType { get; private set; }
+
+        public ConditionalConnectAttribute(Type type) {
+            ConditionalType = type;
+        }
+    }
+}