+++ Editor/Scripts/Drawers/Grouping/FoldoutGroupDrawer.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEditor;
 using RealityProgrammer.OverseerInspector.Runtime.Drawers.Group;
+using RealityProgrammer.OverseerInspector.Editors.Attributes;
 using UnityEditor.AnimatedValues;
 
 namespace RealityProgrammer.OverseerInspector.Editors.Drawers.Group {