+++ Editor/Scripts/Drawers/SeperatorDrawer.cs
@@ -5,6 +5,7 @@ using UnityEngine;
 using UnityEditor;
 using RealityProgrammer.OverseerInspector.Runtime.Drawers;
 using RealityProgrammer.OverseerInspector.Editors.Utility;
+using RealityProgrammer.OverseerInspector.Editors.Attributes;
 
 namespace RealityProgrammer.OverseerInspector.Editors.Drawers {
     [BindDrawerTo(typeof(SeperatorAttribute))]