+++ Runtime/Attributes/Conditional/OverseerConditionalAttribute.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using UnityEngine;
 
 namespace RealityProgrammer.OverseerInspector.Runtime.Validation {
-    public abstract class ConditionalValidationAttribute : BaseOverseerAttribute {
-        public abstract bool Validation(object target);
+    public abstract class OverseerConditionalAttribute : BaseOverseerAttribute {
     }
 }